+++ packages/json-rpc-engine/src/index.test.ts
@@ -6,6 +6,7 @@ describe('@metamask/json-rpc-engine', () => {
       [
         "asV2Middleware",
         "createAsyncMiddleware",
+        "createMethodMiddlewareFactory",
         "createScaffoldMiddleware",
         "getUniqueId",
         "createIdRemapMiddleware",