+++ packages/json-rpc-engine/src/v2/index.test.ts
@@ -15,7 +15,6 @@ describe('@metamask/json-rpc-engine/v2', () => {
         "getUniqueId",
         "isNotification",
         "isRequest",
-        "selectHooks",
       ]
     `);
   });