+++ packages/json-rpc-engine/src/v2/index.ts
@@ -1,6 +1,6 @@
 export { asLegacyMiddleware } from './asLegacyMiddleware';
 export { getUniqueId } from '../getUniqueId';
-export { selectHooks, createMethodMiddleware } from './createMethodMiddleware';
+export { createMethodMiddleware } from './createMethodMiddleware';
 export type { MethodHandler } from './createMethodMiddleware';
 export { createOriginMiddleware } from './createOriginMiddleware';
 export { createScaffoldMiddleware } from './createScaffoldMiddleware';